FSEOF and new partition
